[arcane,std] Déplace 'SplitSDMeshPartitioner' de 'ceapart' vers 'std'.
