+++ arcane/ceapart/src/arcane/cea/srcs.cmake
@@ -2,14 +2,12 @@ set( ARCANE_SOURCES
   BasicRayMeshIntersection.cc
   LimaCutInfosReader.cc
   LimaCutInfosReader.h
-  SplitSDMeshPartitioner.h
   ArcaneCeaVerifierModule.cc
   CartesianConnectivity.h
   CellDirectionMng.h
   FaceDirectionMng.h
   NodeDirectionMng.h
   LimaCutInfosReader.h
-  SplitSDMeshPartitioner.h
   CartesianConnectivity.h
   CartesianMeshRenumberingInfo.h
   CellDirectionMng.h
@@ -21,7 +19,6 @@ set( ARCANE_SOURCES
   )
 
 set(AXL_FILES 
-  SplitSDMeshPartitioner
   ArcaneCeaVerifier
   )
