+++ arcane/src/arcane/std/srcs.cmake
@@ -150,6 +150,7 @@ set(AXL_FILES
   PDESRandomNumberGenerator
   SimpleCsvOutput
   SimpleCsvComparator
+  SplitSDMeshPartitioner
   KdiPostProcessor
   VtkHdfPostProcessor
   VtkHdfV2PostProcessor
